+++ client/src/JobPaginator.jsx
@@ -1,6 +1,8 @@
-import $ from 'jquery';
-
 var React = require('react');
+var PaginatorControls = require('./PaginatorControls');
+var PaginatorList = require('./PaginatorList');
+var PaginatorPreview = require('./PaginatorPreview');
+var PaginatorZoomy = require('./PaginatorZoomy');
 
 class JobPaginator extends React.Component{
     constructor(props) {
@@ -281,205 +283,4 @@ class JobPaginator extends React.Component{
     }
 };
 
-class PaginatorPreview extends React.Component{
-    render = () => {
-        var img = this.props.img
-            ? <img src={this.props.img} alt="" />
-            : '';
-        return (
-            <div className="preview">
-                {img}
-            </div>
-        );
-    }
-};
-
-class PaginatorZoomy extends React.Component{
-    componentDidMount = () => {
-        this.Zoomy.init(document.getElementById('zoomy'));
-        this.componentDidUpdate();
-    }
-
-    componentDidUpdate = () => {
-        this.Zoomy.load(
-            this.props.img,
-            function() {
-                this.Zoomy.resize();
-                this.Zoomy.center();
-                $(this.refs.status).hide();
-            }.bind(this)
-        );
-    }
-
-    render = () => {
-        return (
-            <div>
-                <div ref="status" id="zoomyStatus">Loading...</div>
-                <canvas id="zoomy"></canvas>
-            </div>
-        );
-    }
-};
-
-class PaginatorControls extends React.Component{
-    approveCurrentPageLabel = () => {
-        this.setLabel(this.getLabel(true));
-    }
-
-    getLabel = (useMagic) => {
-        if (typeof useMagic === 'undefined') {
-            useMagic = true;
-        }
-        var label = $(this.refs.labelInput).val();
-        return (label.length === 0 && useMagic)
-            ? this.props.paginator.getLabel(this.props.paginator.state.currentPage)
-            : label;
-    }
-
-    setLabel = (label) => {
-        this.props.paginator.setLabel(this.props.paginator.state.currentPage, label);
-    }
-
-    setLabelPrefix = (str) => {
-        this.setLabel(
-            this.MagicLabeler.replaceLabelPart(this.getLabel(), 'prefix', str, true)
-        );
-    }
-
-    setLabelBody = (str) => {
-        this.setLabel(
-            this.MagicLabeler.replaceLabelPart(this.getLabel(), 'label', str)
-        );
-    }
-
-    setLabelSuffix = (str) => {
-        this.setLabel(
-            this.MagicLabeler.replaceLabelPart(this.getLabel(), 'suffix', str, true)
-        );
-    }
-
-    toggleBrackets = () => {
-        this.setLabel(this.MagicLabeler.toggleBrackets(this.getLabel()));
-    }
-
-    toggleCase = () => {
-        this.setLabel(this.MagicLabeler.toggleCase(this.getLabel()));
-    }
-
-    toggleRoman = () => {
-        var label = this.MagicLabeler.toggleRoman(this.getLabel());
-        if (label === false) {
-            return alert("Roman numeral toggle not supported for this label.");
-        }
-        this.setLabel(label);
-    }
-
-    updateCurrentPageLabel = () => {
-        this.setLabel(this.getLabel(false));
-    }
-
-    render = () => {
-        return (
-            <div className="controls">
-                <div className="group">
-                    <div className="status"></div>
-                    <input type="text" value={this.props.paginator.getLabel(this.props.paginator.state.currentPage, false)} ref="labelInput" id="page" onChange={this.updateCurrentPageLabel} />
-                    <button onClick={this.props.paginator.prevPage}>Prev</button>
-                    <button onClick={function() { this.approveCurrentPageLabel(); this.props.paginator.nextPage(); }.bind(this)}>Next</button>
-                </div>
-                <div className="top">
-                    <ZoomToggleButton paginator={this.props.paginator} />
-                    <button className="primary" onClick={function() { this.approveCurrentPageLabel(); this.props.paginator.save(false); }.bind(this)}>Save</button>
-                    <button className="primary" onClick={function() { this.approveCurrentPageLabel(); this.props.paginator.save(true); }.bind(this)}>Save and Publish</button>
-                </div>
-                <PaginatorControlGroup callback={this.setLabelPrefix} label="prefixes">{this.MagicLabeler.prefixes}</PaginatorControlGroup>
-                <PaginatorControlGroup callback={this.setLabelBody}   label="labels"  >{this.MagicLabeler.labels}</PaginatorControlGroup>
-                <PaginatorControlGroup callback={this.setLabelSuffix} label="suffixes">{this.MagicLabeler.suffixes}</PaginatorControlGroup>
-                <div className="toggles group">
-                    <button onClick={this.toggleBrackets} title="Toggle Brackets">[ ]</button>
-                    <button onClick={this.toggleCase} title="Toggle Case"><i className="fa fa-text-height"></i></button>
-                    <button onClick={this.toggleRoman} title="Toggle Roman Numerals">4<i className="fa fa-fw fa-arrows-h"></i>IV</button>
-                </div>
-                <button onClick={this.props.paginator.autonumberFollowingPages} title="Autonumber Following Pages"><i className="fa fa-sort-numeric-asc"></i></button>
-                <button className="danger" onClick={this.props.paginator.deletePage} title="Delete Current Page"><i className="fa fa-fw fa-trash"></i> Delete Current Page</button>
-            </div>
-        );
-    }
-};
-
-class PaginatorControlGroup extends React.Component{
-    render = () => {
-        var buttons = this.props.children.map(function (item) {
-            var callback = function() {
-                this.props.callback(item);
-            }.bind(this);
-            return (
-                <button onClick={callback} key={item}>{item}</button>
-            );
-        }.bind(this));
-        return (
-            <div className="group" id={this.props.label}>{buttons}</div>
-        );
-    }
-};
-
-class PaginatorList extends React.Component{
-    scrollTo = (thumb) => {
-        var listOffset =
-            this.refs.pageList.offsetTop +
-            (this.refs.thumb0.refs.wrapper.offsetTop - this.refs.pageList.offsetTop);
-        this.refs.pageList.scrollTop = thumb.offsetTop - listOffset;
-    }
-
-    render = () => {
-        var pages = [];
-        for (var i = 0; i < this.props.pageCount; i++) {
-            pages[i] = <Thumbnail ref={"thumb" + i} list={this} selected={i === this.props.paginator.state.currentPage} paginator={this.props.paginator} key={i} number={i} />;
-        };
-        return (
-            <div ref="pageList" className="pageList">{pages}</div>
-        );
-    }
-};
-
-class Thumbnail extends React.Component{
-    selectPage = () => {
-        this.props.paginator.setPage(this.props.number);
-    }
-
-    componentDidUpdate = () => {
-        if (this.props.selected) {
-            this.props.list.scrollTo(this.refs.wrapper);
-        }
-    }
-
-    render = () => {
-        var label = this.props.paginator.getLabel(this.props.number);
-        // check for magic labels:
-        var labelClass = 'label' +
-            (null === this.props.paginator.getLabel(this.props.number, false) ? ' magic' : '');
-        var myClass = 'thumbnail' + (this.props.selected ? ' selected' : '');
-        return (
-            <div onClick={this.selectPage} className={myClass} ref="wrapper">
-              <div className="ratio">
-                <div className="content">
-                  <span className="img-helper"></span>
-                  <img alt="" src={this.props.paginator.getImageUrl(this.props.number, 'thumb')} />
-                </div>
-              </div>
-              <div className="number">{this.props.number + 1}</div>
-              <div className={labelClass}>{label}</div>
-            </div>
-        );
-    }
-};
-
-class ZoomToggleButton extends React.Component{
-    render = () => {
-        return (
-            <button onClick={this.props.paginator.toggleZoom}>{this.props.paginator.state.zoom ? 'Turn Zoom Off' : 'Turn Zoom On'}</button>
-        );
-    }
-};
-
-export default JobPaginator;
+module.exports = JobPaginator;
